Added a basic test and few renamings
